cleaning up files including renaming existing files to standards
